requirements.txt
# Core ML framework dependencies
torch>=2.0.0,<3.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing and validation
pyyaml>=6.0
tqdm>=4.65.0
jsonschema>=4.17.0

# ProjectStrataML framework dependencies
click>=8.1.0
rich>=13.0.0

# System interaction
requests>=2.31.0